extract search action buttons into component and use in list and grid layout
