+++ ui/src/components/routes/hits/search/shared/LayoutSettings.tsx
